pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "how-cli-assist"
version = "1.0.2"
description = "Terminal assistant generating accurate shell commands"
authors = [{ name = "Adem Kouki. Updates by Hypothesis Factory" }]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "google-genai",
    "pyperclip",
    "psutil"
]
requires-python = ">=3.10"

[project.scripts]
how = "how.main:main"
